Opt for bulk save role permissions instead of save on each form change
